+++ tests/Properties.hs
@@ -62,6 +62,7 @@ import QuickCheckUtils
 import Test.Tasty
 import Test.Tasty.QuickCheck
 
+import qualified Properties.ShortByteString as PropSBS
 import qualified Properties.ByteString as PropBS
 import qualified Properties.ByteStringChar8 as PropBS8
 import qualified Properties.ByteStringLazy as PropBL
@@ -620,15 +621,16 @@ explosiveTail = (`L.append` error "Tail of this byte string is undefined!")
 
 testSuite :: TestTree
 testSuite = testGroup "Properties"
-  [ testGroup "StrictWord8" PropBS.tests
-  , testGroup "StrictChar8" PropBS8.tests
-  , testGroup "LazyWord8"   PropBL.tests
-  , testGroup "LazyChar8"   PropBL8.tests
-  , testGroup "Overflow"    overflow_tests
-  , testGroup "Misc"        misc_tests
-  , testGroup "IO"          io_tests
-  , testGroup "Short"       short_tests
-  , testGroup "Strictness"  strictness_checks
+  [ testGroup "ShortByteString" PropSBS.tests
+  , testGroup "StrictWord8"     PropBS.tests
+  , testGroup "StrictChar8"     PropBS8.tests
+  , testGroup "LazyWord8"       PropBL.tests
+  , testGroup "LazyChar8"       PropBL8.tests
+  , testGroup "Overflow"        overflow_tests
+  , testGroup "Misc"            misc_tests
+  , testGroup "IO"              io_tests
+  , testGroup "Short"           short_tests
+  , testGroup "Strictness"      strictness_checks
   ]
 
 io_tests =
@@ -730,6 +732,7 @@ strictness_checks =
         D.take (D.length xs + 1) (D.scanl (+.) '\NUL' (explosiveTail (xs <> D.singleton '\SOH'))) === (D.pack . fmap (D.foldr (+.) '\NUL') . D.inits) xs
     , testProperty "scanl1 is lazy" $ \ xs -> D.length xs > 0 ==> let char1 +. char2 = toEnum (fromEnum char1 + fromEnum char2) in
         D.take (D.length xs) (D.scanl1 (+.) (explosiveTail (xs <> D.singleton '\SOH'))) === (D.pack . fmap (D.foldr1 (+.)) . tail . D.inits) xs
+    , testProperty "unlines is lazy" $ \ xs -> D.take (D.length xs + 1) (D.unlines (xs : error "Tail of this list is undefined!")) === xs `D.snoc` '\n'
     ]
   ]